+++ app/build.gradle
@@ -39,12 +39,12 @@ android {
 dependencies {
     //뷰페이저 추가
     implementation 'androidx.viewpager2:viewpager2:1.0.0'
-
     //retrofit추가
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     // json 변환하기 위한 또라이브러리, gson 예시.
     implementation 'com.google.code.gson:gson:2.8.9'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
+
 
     implementation 'androidx.core:core-ktx:1.9.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
