+++ app/src/main/java/com/example/find_my_matzip/HomeTabActivity.kt
@@ -1,16 +1,15 @@
 package com.example.find_my_matzip
 
-import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.MenuItem
 import androidx.appcompat.app.ActionBarDrawerToggle
 import androidx.fragment.app.Fragment
-import com.example.find_my_matzip.NavTab.NavTabFragment.HomeFragment
-import com.example.find_my_matzip.NavTab.NavTabFragment.MapFragment
-import com.example.find_my_matzip.NavTab.NavTabFragment.MyPageFragment
-import com.example.find_my_matzip.NavTab.NavTabFragment.RankingFragment
-import com.example.find_my_matzip.NavTab.NavTabFragment.SearchReviewFragment
+import com.example.find_my_matzip.navTab.navTabFragment.HomeFragment
+import com.example.find_my_matzip.navTab.navTabFragment.MapFragment
+import com.example.find_my_matzip.navTab.navTabFragment.MyPageFragment
+import com.example.find_my_matzip.navTab.navTabFragment.RankingFragment
+import com.example.find_my_matzip.navTab.navTabFragment.SearchReviewFragment
 import com.example.find_my_matzip.databinding.ActivityHomeTabBinding
 
 class HomeTabActivity : AppCompatActivity() {
