+++ app/src/main/java/com/example/find_my_matzip/navTab/navTabFragment/MapFragment.kt
@@ -1,21 +1,20 @@
-package com.example.find_my_matzip.NavTab.NavTabFragment
+package com.example.find_my_matzip.navTab.navTabFragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.find_my_matzip.R
-import com.example.find_my_matzip.databinding.FragmentHomeBinding
 import com.example.find_my_matzip.databinding.FragmentMapBinding
 
 
 
-class HomeFragment : Fragment() {
-    lateinit var binding: FragmentHomeBinding
+
+class MapFragment : Fragment() {
+    lateinit var binding: FragmentMapBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = FragmentHomeBinding.inflate(layoutInflater)
+        binding = FragmentMapBinding.inflate(layoutInflater)
 
 
     }
@@ -24,7 +23,7 @@ class HomeFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentHomeBinding.inflate(layoutInflater, container, false)
+        binding = FragmentMapBinding.inflate(layoutInflater, container, false)
         return binding.root
     }
 }
