+++ app/src/main/java/com/example/find_my_matzip/navTab/navTabFragment/SearchReviewFragment.kt
@@ -1,12 +1,10 @@
-package com.example.find_my_matzip.NavTab.NavTabFragment
+package com.example.find_my_matzip.navTab.navTabFragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.find_my_matzip.R
-import com.example.find_my_matzip.databinding.FragmentMatZalalBinding
 import com.example.find_my_matzip.databinding.FragmentSearchReviewBinding
