+++ app/src/main/res/layout/fragment_child_home1.xml
@@ -3,12 +3,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".navTab.navTabFragment.MatZalalFragment">
+    tools:context=".navTab.navTabFragment.homeFragChildFrags.ChildHomeFragment1">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="맛잘알1" />
+        android:text="@string/hello_blank_fragment" />
 
 </FrameLayout>
