+++ app/src/main/res/layout/fragment_home.xml
@@ -3,11 +3,33 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NavTab.NavTabFragment.HomeFragment">
+    tools:context=".navTab.navTabFragment.HomeFragment">
 
-    <TextView
+    <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="HomeFragment" />
+        android:orientation="vertical">
+
+        <!--텍스트뷰, 리사이클러뷰 붙이기-->
+        <TextView
+            android:id="@+id/firsttext"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="HomeFragment" />
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="메인화면 리사이클러뷰"
+            android:textSize="20dp"
+            android:gravity="center"
+            ></TextView>
+
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/retrofitRecyclerView2"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            ></androidx.recyclerview.widget.RecyclerView>
+
+    </LinearLayout>
 
 </FrameLayout>
