+++ app/src/main/res/layout/fragment_map.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NavTab.NavTabFragment.MapFragment">
+    tools:context=".navTab.navTabFragment.MapFragment">
 
     <TextView
         android:layout_width="match_parent"
