+++ app/src/main/res/layout/fragment_ranking.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NavTab.NavTabFragment.RankingFragment">
+    tools:context=".navTab.navTabFragment.RankingFragment">
 
     <TextView
         android:layout_width="match_parent"
