+++ app/src/main/res/layout/fragment_search_review.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NavTab.NavTabFragment.SearchReviewFragment">
+    tools:context=".navTab.navTabFragment.SearchReviewFragment">
 
     <TextView
         android:layout_width="match_parent"
