+++ app/src/main/res/layout/home_fragment_item.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:orientation="vertical"
+    android:padding="15dp"
+    >
+
+    <ImageView
+        android:id="@+id/retrofitProfileImg2"
+        android:layout_width="match_parent"
+        android:layout_height="300dp"
+        android:src="@drawable/burger"
+        android:scaleType="centerCrop"
+        ></ImageView>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:padding="5dp"
+        android:layout_gravity="center"
+        >
+        <TextView
+            android:id="@+id/retrofitTitleView"
+            android:textSize="15sp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="title"
+            ></TextView>
+        <TextView
+            android:layout_marginTop="10dp"
+            android:id="@+id/retrofitContentView"
+            android:textSize="15sp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="content"
+            ></TextView>
+    </LinearLayout>
+
+</LinearLayout>
